[BEAM-5356] Move metrics related functionality to common place
