+++ sdks/java/testing/load-tests/src/main/java/org/apache/beam/sdk/loadtests/GroupByKeyLoadTest.java
@@ -30,16 +30,14 @@ import org.apache.beam.sdk.io.synthetic.SyntheticBoundedIO;
 import org.apache.beam.sdk.io.synthetic.SyntheticBoundedIO.SyntheticSourceOptions;
 import org.apache.beam.sdk.io.synthetic.SyntheticOptions;
 import org.apache.beam.sdk.io.synthetic.SyntheticStep;
-import org.apache.beam.sdk.metrics.Counter;
-import org.apache.beam.sdk.metrics.Distribution;
-import org.apache.beam.sdk.metrics.Metrics;
+import org.apache.beam.sdk.loadtests.metrics.MetricsMonitor;
+import org.apache.beam.sdk.loadtests.metrics.MetricsPublisher;
 import org.apache.beam.sdk.options.ApplicationNameOptions;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.options.Validation;
-import org.apache.beam.sdk.testutils.metrics.MetricsReader;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.GroupByKey;
 import org.apache.beam.sdk.transforms.ParDo;
@@ -125,7 +123,7 @@ public class GroupByKeyLoadTest {
 
     for (int branch = 0; branch < options.getFanout(); branch++) {
       applySyntheticStep(input, branch, syntheticStep)
-          .apply(ParDo.of(new Monitor()))
+          .apply(ParDo.of(new MetricsMonitor("gbk")))
           .apply(format("Group by key (%s)", branch), GroupByKey.create())
           .apply(
               format("Ungroup and reiterate (%s)", branch),
@@ -135,18 +133,7 @@ public class GroupByKeyLoadTest {
     PipelineResult result = pipeline.run();
     result.waitUntilFinish();
 
-    printMetrics(result);
-  }
-
-  private static void printMetrics(PipelineResult result) {
-    MetricsReader resultMetrics = new MetricsReader(result, "gbk");
-
-    long totalBytes = resultMetrics.getCounterMetric("totalBytes.count", -1);
-    long startTime = resultMetrics.getStartTimeMetric(System.currentTimeMillis(), "runtime");
-    long endTime = resultMetrics.getEndTimeMetric(System.currentTimeMillis(), "runtime");
-
-    System.out.println(String.format("Total bytes: %s", totalBytes));
-    System.out.println(String.format("Total time (millis): %s", endTime - startTime));
+    MetricsPublisher.toConsole(result, "gbk");
   }
 
   private static PCollection<KV<byte[], byte[]>> applySyntheticStep(
@@ -190,21 +177,6 @@ public class GroupByKeyLoadTest {
           }
         }
       }
-    }
-  }
-
-  private static class Monitor extends DoFn<KV<byte[], byte[]>, KV<byte[], byte[]>> {
-
-    private static final String NAMESPACE = "gbk";
-
-    private final Distribution timeDistribution = Metrics.distribution(NAMESPACE, "runtime");
-    private final Counter totalBytes = Metrics.counter(NAMESPACE, "totalBytes.count");
-
-    @ProcessElement
-    public void processElement(ProcessContext c) {
-      timeDistribution.update(System.currentTimeMillis());
-      totalBytes.inc(c.element().getKey().length + c.element().getValue().length);
-      c.output(c.element());
     }
   }
 }
